fix(server): #920 — move the formula guard above both spreadsheet writers
`cargo check -p fraiseql-server --no-default-features --features rest,export-xlsx`
did not compile: the XLSX writer imported `guard_formula_injection` from
`streaming::csv`, but `export-csv` and `export-xlsx` are independently
selectable and `pub mod csv` is gated on the former. An operator who wanted
spreadsheet export without CSV got `E0432: could not find csv in super`, and the
only workaround was enabling the response format they had excluded.

## Changes

- `guard_formula_injection` + `FORMULA_INJECTION_SENTINELS` move to `streaming/mod.rs`
  under `cfg(any(export-csv, export-xlsx))`; both writers import from there. The
  threat model moves with the function rather than staying in `value_to_csv_field`.
- Its 8 tests move to `streaming::tests::formula_injection` under the same cfg, so
  an XLSX-only build compiles and runs them too.
- Two new feature-matrix combos, `server-rest-export-xlsx` and
  `server-rest-export-csv` — this was a P01-class hole, no leg built either writer
  alone. One per writer so a cross-module import in either direction reddens CI.
  `export-parquet` gets no combo: it has zero `cfg` usage, so building it exercises
  nothing (that ghost is #1012's to resolve).

## Verification

✅ All four export combinations compile: `rest,export-xlsx` / `rest,export-csv` /
   both / `rest,export-parquet`
✅ `cargo test -p fraiseql-server --lib --all-features` — 1848 passed; the 8 guard
   tests execute by name under `streaming::tests::formula_injection`
✅ `cargo clippy -p fraiseql-server --all-targets --all-features -- -D warnings`
✅ `go build ./...` in `.dagger`
✅ RED-proven: restoring `use super::csv::guard_formula_injection` makes the new
   combo's exact command exit 101 with E0432

Closes #920
