+++ crates/fraiseql-server/src/routes/rest/streaming/csv/mod.rs
@@ -31,10 +31,13 @@ use fraiseql_core::{
 };
 use futures::stream;
 
-use super::super::{
-    export_config::ExportConfig,
-    handler::{PreferHeader, ResolvedGetQuery, RestError, RestHandler, set_request_id},
-    params::PaginationParams,
+use super::{
+    super::{
+        export_config::ExportConfig,
+        handler::{PreferHeader, ResolvedGetQuery, RestError, RestHandler, set_request_id},
+        params::PaginationParams,
+    },
+    guard_formula_injection,
 };
 
 /// Content type for CSV responses.
@@ -443,30 +446,6 @@ fn value_to_csv_field(v: &serde_json::Value) -> String {
         serde_json::Value::Number(n) => guard_formula_injection(&n.to_string()),
         serde_json::Value::String(s) => guard_formula_injection(s),
         other => guard_formula_injection(&serde_json::to_string(other).unwrap_or_default()),
-    }
-}
-
-/// Single-byte sentinels that trigger formula evaluation in Excel /
-/// `LibreOffice` / Numbers when they appear as the first character of a
-/// cell.  Tab and CR are included because Excel will treat them as
-/// whitespace-prefixed formula starters when followed by `=` etc., and
-/// because both are present in OWASP's reference list for this attack.
-const FORMULA_INJECTION_SENTINELS: [char; 6] = ['=', '+', '-', '@', '\t', '\r'];
-
-/// Prefixes `value` with a single quote when its first character would
-/// otherwise be interpreted by a spreadsheet application as the start of
-/// a formula.  See the `value_to_csv_field` docstring for the threat
-/// model.  Returns `value` unchanged for non-dangerous prefixes (the
-/// common case) so the function is allocation-free on the hot path.
-pub(crate) fn guard_formula_injection(value: &str) -> String {
-    match value.chars().next() {
-        Some(c) if FORMULA_INJECTION_SENTINELS.contains(&c) => {
-            let mut out = String::with_capacity(value.len() + 1);
-            out.push('\'');
-            out.push_str(value);
-            out
-        },
-        _ => value.to_owned(),
     }
 }
